feat: add unit test job to CI configuration and create MovieCardPreview component
